refactor: IDCLEntity extraction and cleanup
- Moved IDCLEntity to a new file

- Had to replace the DCLName and DCLTransform model references with object. This is because the interface cannot reference system implementations. In any case, this makes a design flaw evident - the events that were referencing specific components will have to be moved elsewhere in the future.
